+++ changelog.md
@@ -1,5 +1,7 @@
 ## additions and changes
 
+- Added pattern matching implementation as described in
+  [nim-lang/RFCs#245](https://github.com/nim-lang/RFCs/issues/245)
 - Added module `astdsl` that contains macro `buildAst`. That is a DSL for convenient
   construction of Nim ASTs.
 - Added module `pointers` containing `toUncheckedArray`.
@@ -11,3 +13,4 @@
 
 
 - Added `jsxmlhttprequest` module for [`XMLHttpRequest`](https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest) for the JavaScript target.
+- Added `jsheaders` module for [`Headers`](https://developer.mozilla.org/en-US/docs/Web/API/Headers) for the JavaScript target.